Make a usable version of getEventByCorrelationId (#19)
This adds @baronfel's Newtonsoft.Json.FSharp.Idiomatic to provide custom JsonConverter that are more friendly for F# serializing.  Now that the json that gets serialized turns out to be sane, we can use raw sql to query the EventRead table easily.  Although this doesn't use Martens Linq generation, it's the quickest option to getting a production ready version